unifying table generation structure for different number systems
