+++ education/tables/fixpnts.cpp
@@ -1,70 +1,16 @@
 // fixpnts.cpp: generates encoding tables of fixed-point configurations
 //
-// Copyright (C) 2017-2021 Stillwater Supercomputing, Inc.
+// Copyright (C) 2017-2022 Stillwater Supercomputing, Inc.
 //
 // This file is part of the universal numbers project, which is released under an MIT Open Source license.
-#if defined(_MSC_VER)
-#pragma warning(disable : 4514)  // unreferenced function is removed
-#pragma warning(disable : 4710)  // function is not inlined
-#pragma warning(disable : 5045)  // compiler will insert Spectre mitigation for memory load if /Qspectre switch specified
-#endif 
+#include <universal/utility/directives.hpp>
 // Configure the fixpnt template environment
 // first: enable general or specialized fixed-point configurations
 #define FIXPNT_FAST_SPECIALIZATION
 // second: enable/disable fixpnt arithmetic exceptions
 #define FIXPNT_THROW_ARITHMETIC_EXCEPTION 1
-
 #include <universal/number/fixpnt/fixpnt.hpp>
-#include <universal/verification/fixpnt_test_suite.hpp>
-
-// generate a full binary representation table for a given posit configuration
-template<size_t nbits, size_t rbits>
-void GenerateFixedPointTable(std::ostream& ostr, bool csvFormat = false) {
-	const size_t size = (1 << nbits);
-	sw::universal::fixpnt<nbits, rbits>	p;
-	if (csvFormat) {
-		ostr << "\"Generate Fixed-Point Lookup table for a FIXPNT<" << nbits << "," << rbits << "> in CSV format\"" << std::endl;
-		ostr << "#, Binary, sign, scale, value\n";
-		for (size_t i = 0; i < size; i++) {
-			p.setbits(i);
-			ostr << i << ","
-				<< to_binary(p) << ","
-				<< p.sign() << ","
-				<< scale(p) << ","
-				<< p
-				<< '\n';
-		}
-		ostr << std::endl;
-	}
-	else {
-		ostr << "Generate Fixed-Point Lookup table for a FIXPNT<" << nbits << "," << rbits << "> in TXT format" << std::endl;
-
-		const size_t index_column = 5;
-		const size_t bin_column = 16;
-		const size_t sign_column = 8;
-		const size_t scale_column = 8;
-		const size_t value_column = 30;
-		const size_t format_column = 16;
-
-		ostr << std::setw(index_column) << " # "
-			<< std::setw(bin_column) << "Binary"
-			<< std::setw(sign_column) << "sign"
-			<< std::setw(scale_column) << "scale"
-			<< std::setw(value_column) << "value"
-			<< std::setw(format_column) << "format"
-			<< std::endl;
-		for (size_t i = 0; i < size; i++) {
-			p.setbits(i);
-			ostr << std::setw(4) << i << ": "
-				<< std::setw(bin_column) << to_binary(p)
-				<< std::setw(sign_column) << p.sign()
-				<< std::setw(scale_column) << scale(p)
-				<< std::setw(value_column) << p << " "
-				<< std::setw(format_column) << std::right << p
-				<< std::endl;
-		}
-	}
-}
+#include <universal/number/fixpnt/table.hpp>
 
 int main(int argc, char** argv)
 try {
