+++ include/universal/number/posit/table.hpp
@@ -1,14 +1,9 @@
 #pragma once
-// table.hpp: generate a posit table
+// table.hpp: generate a table of encoding and values for fixed-size arbitrary configuration posits
 //
-// Copyright (C) 2017-2021 Stillwater Supercomputing, Inc.
+// Copyright (C) 2017-2022 Stillwater Supercomputing, Inc.
 //
 // This file is part of the universal numbers project, which is released under an MIT Open Source license.
-
-#include <iostream>
-#include <iomanip>
-#include <cmath>  // for frexp/frexpf
-#include <typeinfo>  // for typeid()
 
 namespace sw { namespace universal {
